`refactor: ロード/ダウンロードフローとABC出力をmain.tsから分離` ## 概要 `main.ts` に集中していた入出力まわりの処理を分離し、責務を明確化しました。 あわせて、MusicXML→ABC変換ロジックを `abc-io.ts` に集約しました。
## 変更内容

### 1. ロード処理の分離
- `src/ts/load-flow.ts` を新規追加
- 入力種別（新規作成 / MusicXML / ABC）と入力モード（ファイル / テキスト）を解決する処理を `resolveLoadFlow` に移動
- エラーは `LoadFlowResult`（Result型）で返す構成に変更

### 2. ダウンロード処理の分離
- `src/ts/download-flow.ts` を新規追加
- 以下を `main.ts` から移動
  - MusicXMLダウンロードペイロード生成
  - MIDIダウンロードペイロード生成
  - ABCダウンロードペイロード生成
  - 共通ダウンロード実行（`triggerFileDownload`）

### 3. ABC出力ロジックの集約
- `src/ts/abc-io.ts` に `exportMusicXmlDomToAbc(doc: Document): string` を追加
- `main.ts` 内にあった `convertMusicXmlToAbc` 実装を削除し、`abc-io.ts` へ移管
- 既存の `AbcCompatParser` と同じABC入出力モジュールに統一

### 4. main.ts の整理
- `onLoadClick` / `onDownload*` の実装を簡素化
- 処理本体は `load-flow.ts` / `download-flow.ts` / `abc-io.ts` に委譲

### 5. 生成物の更新
- 変更に伴い `src/js/main.js` と `mikuscore.html` / `mikuscore-src.html` を更新

## 変更ファイル
- `src/ts/load-flow.ts` (new)
- `src/ts/download-flow.ts` (new)
- `src/ts/abc-io.ts`
- `src/ts/main.ts`
- `src/js/main.js`
- `mikuscore-src.html`
- `mikuscore.html`

## 期待効果
- `main.ts` の責務分離と可読性向上
- ロード/ダウンロード仕様のテスト・保守容易性向上
- ABC I/O関連ロジックの配置一貫性向上

## 備考
- 機能追加というより、処理移管と構造改善が中心です。
